Make branch links preserve tree contents
When detecting and creating branch roots, there are two cases:
 1) It is a directoy copy from the toplevel of another branch to create
    the top-level of the current branch. In this case, no fileops are
    generated because they would do nothing in git.
 2) It is not, and the full fileop list is kept.

In the first case, when the branch root is connected to the branch
point, the tree of the latter should be inherited by the former. This is
the normal behavior of branch creation in git, and is already handled
correctly by reposurgeon.

In the second case, the contents of the new branch tree are given by the
file operations that have been kept. Sometimes the copy operation
expressed by all these fileops don't amount to a full copy of a branch
dir to another. For instance it can be a subtree extraction [1], or even
enough separate file copies to trigger the branching heuristics. Then
the tree of the computed branch point shouldn't be inherited at all
since that would pollute the branch contents. This case wasn't correctly
handled by reposurgeon.

To ensure the new branch starts with a fresh tree if needed, prepend a
deleteall to the first non-trivial [2] commit of the branch, only in
case 2, and only when the branch root is actually linked back to the
branch point to avoid prepending unnecessary deleteall operations to a
lot of branches. In other words, only prepend a deleteall if the result
would be wrong without.

Also update the testsuite to record the new correct behavior.

[1] using a subdir of a branch to make that the new branch top directory
[2] we consider it non-trivial when it is non empty, or has more than
    one child. This gives the correct result while avoiding to put a
    file operation in a commit that would otherwise have been tagified.