Try harder to find the exact parent of a merge
When a svn merge is performed and the mergeinfo populated, no effort is
made on the part of svn to ensure that the end revision really is a
revision touching the source branch. While this isn't important in
Subversion where there are no branches but rather directories, this
makes merges generated by Reposurgeon completely wrong: the merge
suddenly isn't between the correct branches anymore !

Instead of blindly trusting svn:mergeinfo, search backwards for the last
commit on the source branch. This method has a notable side-effect: it
can choose the right split commit of a revision touching multiple
branches. No more "cannot resolve from split commit" error.