Skip commits with no branch in last_relevant_commit
Ensure that the startswith test is only made against non-empty strings
(or else it trivially matches, possibly masking another commit that
really was on the branch). This fixes some lookbacks in the branchlink
detection code when there are commits on the "root" branch).

In particular the references test needs adjustment.