Fix branchlink detection for full branch copies
The branchlink detection code considered all copies in order to find
whicn revision to link back from; in particular it could choose a single
file import from another revision (or another branch altogether) if it
happened to be later. But most of the files in the tree probably come
from the directory copy, so that revision is more relevant.

Moreover, not linking to the commit for the source dir copy can be
flatly wrong if it was detected as a branch copy and the directory copy
didn't get a corresponding fileop.

Fix the logic to select the directory copy if it exists.

Note that in some cases this means that the detected link is less
interesting than what was before: when doing an "svn copy" from a
directory in an SVN worktree, SVN doesn't generate a single dir-add from
the previous revision but a dir-add from the revision that created the
dir then populates the directory with file copies from their last
respective revision. This comes from the fact that SVN tracks a
different revision per-file.

The linkback previously was to the latest from_rev in the set, which
corresponded to what the user asked of subversion at the time. Now it
will be to the from_rev of the source directory, which can be a lot
earlier but short of computing manifests there is no mean to tell that
case apart from a directory copy followed by unrelated file copies where
the previous behavior is wrong.