Don't use Tag and Reset as set members (attachments)
Since the hash of an object without custom __hash__ is based on its
memory location, it is not stable which can cause sets and dicts to
search for the object in the wrong hash table slot and raise KeyError
even if the object is in it. Pypy might even outright crash.

Make Commit.attachments a list to sidestep the problem. The same
structure is already used for Commit.parent_nodes and Commit.child_nodes
without causing much slowdown even with linear search and deletion.