Add a test checking branch/tag saving during delete
The current delete logic can sometimes clobber the branch of the parent
or child of a deleted commit. This means that only one lightweight tag
of a deleted series will be kept, and worse that branch tips can become
misplaced as a side-effect or even branches completely forgotten.

Record that current behaviour in a test.