Mark most classes as not hashable
By setting their __hash__ member to None, we tell Python not to generate
a default hash function for them. This prevents their use as set members
or dict keys with an unstable hash in the future. This is particularly
important for Reset, Tag, and Commit, but is done for all classes that
are instanciated enough to be candidates for future problems.