Make the ? operator work in inner expressions
Contrary to what told the documentation, the "?" operator was only
accepted at the end of the whole selection set expression but triggered
a "trailing junk" error if used in {...} subexpressions. Ensure "?" is
accepted in all cases by making eval_term() call eval_expression()
instead of eval_disjunct() for subexpression parsing.

This changes the parser.chk results for expressions which were
previously considered erroneous.