Ensure whitespace is preserved in tests
Git normally considers trailing whitespace at EOL or blank lines at EOF
to be mistakes, and with the "apply.whitespace" configuration set to
"fix" Git fixes these mistakes when applying, cherry-picking or rebasing
commits. Most of the times this is desirable behavior, but spaces (and
even more empty lines) are significant in reposurgeon tests.

Add a .gitattributes file to ensure that Git ignores all whitespace in
the test/ directory.