combine(): only move fileops before the first parent
When used with the --pushback option, combine() ensures that the fileops
that are pushed to parent commits still have their blob appearing before
use in the stream. It did so by moving all relevant blobs before the
parent with lowest index.

6792261dee218be223cf1e72403c16461bcb8b66 made combine() only push
fileops back to the first (or primary) parent because other parents do
not take part in fileop inheritance. In particular, the relevant blobs
need no longer be moved before the lowest index parent, just before the
primary parent.

Do so, to avoid unneeded movement in the stream.