Fix tagify_empty() when |commits| is not None
tagify_empty() expected an iterable of event objects as its commits
argument and used enumerate() on it to get the corresponding indices and
avoid calls to event.index(). Because enumerate() gives the rank in the
iterable rather than in self.events, the thus computed indices are wrong
unless self.events itself is passed as the commits argument.

Make the function expect indices instead, as delete() does, and fetch
the corresponding events when needed. Also update the only caller with a
non-trivial commits argument.