do_reset: Fix collision detection
The code checked if there were a Reset not already in the resets list yet
with ref equal to that of resets[0], and refused to perform the move in
that case. But the resets list was populated with all Reset events whose
ref was resetname. The collision test thus checks if there is a Reset
referencing resetname which hasn't been added to the resets list, but that
cannot happen.

The intent was instead to check if there were Reset events that already
referenced the new name chosen by the user. Fix the test to do that, and
use the cleaner and faster any() builtin for that purpose.