Fix @min() and @max() functions
The function handlers used the Python min()/max() builtins with argument
unpacking, e.g. "min(*subargs)". Because subarg is a set, thus iterable,
min() then gets passed all indices in the selection set as separate
arguments. This is correct (though non-optimal) most of the times, but
breaks when subarg has only one element. Indeed, the behaviour of min()
changes with a single argument: it then expects an iterable and returns
the minimum yielded value. In this case because its sole argument is a
(non-iterable) int, Python raises a TypeError.

Instead, always rely on that alternate behaviour by not unpacking subarg
at all. Instead of subarg being iterated when building the argument
list, it will be iterated by min directly, which gives the correct
result regardless of whether the set is a singleton or not. It is even
faster because Python won't have to keep track of a potentially huge
argument list.

Also document the fact that because the starting "(" is not removed by
eval_funcall(), the trailing ")" will be eaten by eval_expression().