In svn conversions, discard .gitignore files in the Subversion repository.
The old behavior of having them randomly step on the stuff generated from
svn:ignore properties proved to be unacceptable.  The alternative to
discarding them with a warning would have been some attempt at merge
logic guaranteed to be a sertious defect attractor.

Because such files tend to be simple, duplicative of svn:ignore
properties, and chane seldom, punting this to be done by hand
seems reasonable.